refactor: rename package from lib_stringmatcher to stringmatcher
