+++ lib-stringmatcher/src/main/java/com/example/stringmatcher/StringMatcher.kt
@@ -1,4 +1,4 @@
-package com.example.lib_stringmatcher
+package com.example.stringmatcher
 
 interface StringMatcher {
     fun isMatched(
